+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-- Add log_dir configuration in mprocs.yaml
+- Add proc_log and procs.*.log configuration in mprocs.yaml
 
 ## 0.8.3 - 2026-01-21